Added functionality to have a Coop New Game button on the main menu.
In addition, the page opened by the button is a copy of the New Game
page currently. Clicking any of the game difficulties will effectively
start a new singleplayer game, but it uses a slightly different coop
console command, so the code may break somewhere that is unclear even
after play-testing the beginning and shooting the first NPCs with no
hiccups.

The added gamedata folder is important as you will need to replace your
repective gamedata files with the ones added here. Only the changed
gamedata files have been added here. All other gamedata files to get
this build to work should be the same as what is extracted from the
Stalker Call of Pripyat installation.